fix some compilation issues

+++ src/cmplayer/widgets.hpp
@@ -102,7 +102,7 @@ public:
 	static void retranslate(QComboBox *combo) {
 		const auto items = EnumInfo<Enum>::items();
 		Q_ASSERT(items.size() == combo->count());
-		for (int i=0; i<items.size(); ++i)
+		for (int i=0; i<(int)items.size(); ++i)
 			combo->setItemText(i, EnumInfo<Enum>::description(items[i].value));
 	}
 	Enum currentValue() const { return EnumInfo<Enum>::from(currentData().toInt()); }
